♻ refactor: API 훅 수정
