--- src/api/axiosConfig.js
@@ -1,47 +1,0 @@
-import axios from 'axios'
-
-// axios 인스턴스 생성
-const api = axios.create({
-  baseURL: import.meta.env.VITE_API_BASE_URL || 'http://localhost:8080',
-  timeout: 30000,
-  headers: {
-    'Content-Type': 'application/json',
-  },
-})
-
-// 요청 인터셉터 - 모든 요청에 토큰 추가
-api.interceptors.request.use(
-  (config) => {
-    // localStorage에서 토큰 가져오기
-    const token = localStorage.getItem('access-token')
-
-    if (token) {
-      // Authorization 헤더에 Bearer 토큰 추가
-      config.headers.Authorization = `Bearer ${token}`
-    }
-
-    return config
-  },
-  (error) => {
-    return Promise.reject(error)
-  },
-)
-
-// 응답 인터셉터 - 에러 처리
-api.interceptors.response.use(
-  (response) => {
-    return response
-  },
-  (error) => {
-    // 401 Unauthorized 에러 처리
-    if (error.response && error.response.status === 401) {
-      console.error('인증 에러: 토큰이 만료되었거나 유효하지 않습니다.')
-      // 필요시 로그인 페이지로 리다이렉트
-      // router.push('/login');
-    }
-
-    return Promise.reject(error)
-  },
-)
-
-export default api
